Add desktop POST file download and fix iOS download path
