--- src/components/Search/SearchActionOptionsUtils.desktop.tsx
@@ -1,8 +1,0 @@
-import type {DropdownOption} from '@components/ButtonWithDropdownMenu/types';
-import type {SearchHeaderOptionValue} from './SearchPageHeader';
-
-function getDownloadOption(): DropdownOption<SearchHeaderOptionValue> | undefined {
-    return undefined;
-}
-
-export default getDownloadOption;
